fix last_seen bug
